separate out List type into header.hpp
